Front y back actualizado a la nueva base de datos

+++ GestionResiduos-Backend/src/main/java/com/IntegradorII/GestionResiduos/repository/ReporteRepository.java
@@ -3,6 +3,5 @@ package com.IntegradorII.GestionResiduos.repository;
 import com.IntegradorII.GestionResiduos.entity.Reporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReporteRepository extends JpaRepository<Reporte, Long> {
-
+public interface ReporteRepository extends JpaRepository<Reporte, Integer> {
 }
